requirements.txt
appdirs==1.4.4
asgiref==3.4.1
beautifulsoup4==4.9.3
black==21.7b0
CacheControl==0.12.10
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
configparser==5.2.0
coolsms-python-sdk==2.0.3
cryptography==3.4.7
defusedxml==0.7.1
Django==3.0.6
django-allauth==0.45.0
firebase-admin==5.2.0
google-api-core==1.31.1
google-api-python-client==2.15.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.5
google-cloud-core==2.2.2
google-cloud-firestore==2.3.4
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.3.1
googleapis-common-protos==1.53.0
grpcio==1.44.0
httplib2==0.19.1
idna==3.2
lxml==4.6.3
mailchimp-marketing==3.0.74
msgpack==1.0.3
mypy-extensions==0.4.3
oauth2client==4.1.3
oauthlib==3.1.1
packaging==21.0
pathspec==0.9.0
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.12.0
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
slack-sdk==3.8.0
soupsieve==2.2.1
sqlparse==0.4.1
tomli==1.2.1
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.26.6
user-agents==2.2.0
